Use module/package/class dependencies to order Verilog source files
This is extremely heuristic and not based on a careful reading of
the LRM, I just need this to Work For My Project.

Note that we don't actually need to order module dependencies at all,
instantiating modules before declaring them is extremely okay.
But we try keeping things in module hierarchy order anyways as a
heuristic in case we missed some other important decl in the AST
whose order does matter...
Of course if the design has both a real package dependency one way
between two files, and a module dependency the other way,
we risk giving an incorrect order (basically at random).
But it'll work for me, so let's deal with that small problem Later™ =]

This parser lib is also a little strict (doesn't like trailing commas,
or comments on the same line as a `define), but it's not *too* bad.